+++ Sources/SuperwallKit/Permissions/PermissionType.swift
@@ -16,4 +16,5 @@ enum PermissionType: String, Decodable {
   case readImages = "read_images"
   case contacts
   case camera
+  case tracking
 }